AbstractOddFrame

OddFrames are a name given to any type that is a sub-type of AbstractOddFrame. AbstractOddFrames must all hold the following fields:

labels::Array/Tuple{Symbol}

columns::Array/Tuple{Any}

types::Array/Tuple{Type}

head::Function

dtype::Function

not::Function

only::Function

describe::Function

Missing docstring.

Missing docstring for AbstractOddFrame. Check Documenter's build log for details.

AbstractMutableOddFrame

AbstractMutableOddFrame is similar to AbstractOddFrame, only requires that both the OddFrame and its values are mutable. Along with this comes these additional features, some of which are shared with AbstractOddFrame

labels::Array{Symbol}

columns::Array{Any}

types::Array

head::Function

dtype::Function

not::Function

only::Function

describe::Function

drop!::Function

dtype!::Function

merge!::Function

only!::Function

apply!::Function

fill!::Function

Missing docstring.

Missing docstring for OddFrames.AbstractMutableOddFrame. Check Documenter's build log for details.

OddFrame

Missing docstring.

Missing docstring for OddFrame. Check Documenter's build log for details.

ImmutableOddFrame

Missing docstring.

Missing docstring for ImmutableOddFrame. Check Documenter's build log for details.

Indexing

Indexing in Oddframes.jl has a few key consistencies that are universal in the package and important to note. The first of these is the different types you can index with. The operation of the indexing is going to change depending on what type we are providing. The first way to index an OddFrame is by using a symbol. This will index the labels, and return the corresponding column of data.

Indexing columns

Base.getindex — Method

Interface

Indexing

getindex(od::AbstractOddFrame, col::Symbol) -> ::Array{Any}

Gets the corresponding index of columns from col's position in labels. Returns a 1-D array of the column's data.

example

od = OddFrame(:A => [5, 10, 15])
od[:A]
[5, 10, 15]

The same functionality can also be achieved with a string.

Missing docstring.

Missing docstring for getindex(::AbstractOddFrame, ::Symbol). Check Documenter's build log for details.

We can also use the at key-word argument, which always denotes that we would like to call a column. Otherwise, indexing with an integer will always call the observation. We can also index in this same way with a range.

Base.getindex — Method

Interface

Indexing

getindex(od::AbstractOddFrame, axis::UnitRange) -> ::Vector{Array{any}}

Indexing Returns all columns within range at.

example

od = OddFrame(:A => [5, 10, 15], :B => [5, 10, 15])
od[1:2]
[[5, 10, 15], [5, 10, 15]]

Missing docstring.

Missing docstring for getindex(::AbstractOddFrame; at::Int64). Check Documenter's build log for details.

Indexing rows

Indexing full rows can be done by simply applying the same indexing but replacing the key-word argument at with a positional argument.

Base.getindex — Method

Interface

Indexing

getindex(od::AbstractOddFrame, observation::Int64, at = 1) -> ::Any

Used to select an observation observation from column at on od.

example

od = OddFrame(:A => [5, 10, 15], :B => [8, 12, 10])
od[1, at = 1]

5

Missing docstring.

Missing docstring for getindex(::AbstractOddFrame, ::UnitRange). Check Documenter's build log for details.

Indexing observations

Row indexes are done with three different types, the BitArray, Integers, and UnitRanges. The latter two can be done using the same indexing calls. We use the at key-word here to denote which column we would like to call.

Missing docstring.

Missing docstring for getindex(::AbstractOddFrame, ::Int64; at::Any). Check Documenter's build log for details.

Base.getindex — Method

Interface

Indexing

getindex(od::AbstractOddFrame, mask::BitArray) -> ::OddFrame

Drops values when mask's equivalent index is false. Can be used to filter data. found on the OddFrame. Does not mutate the data, provides a return.

example

od = OddFrame(:A => [5, 10, 15])
od[od[:A] .== 5]
columns(od)

[5]

Missing docstring.

Missing docstring for getindex(::AbstractOddFrame, ::UnitRange). Check Documenter's build log for details.

These same indexing ideas carry into the member functions and the rest of the module, so it is certainly something to pay attention to.

Iteration

Iteration using OddFrames.jl is also incredibly simple. There are a few ways we can iterate an OddFrame. By default, iterating over the OddFrame will loop the columns of the OddFrame.

Base.iterate — Method

Interface

Iteration

iterate(fg::OddFrameContainer, axis::Int64) -> ::Any

Partitions OddFrame columns to be iterated.

example

od = OddFrame(:A => [5, 10, 15])
od2 = OddFrame(:B => [5, 10, 11])
fg = FrameGroup(od, od2)

fg[1]

[5, 10, 15]

However, in order to iterate names, or explicitly call the columns to be iterated, you can use labels() and columns() respectively.

OddFrames.columns — Method

Interface

Iteration

columns(od::AbstractOddFrame) -> ::Array{AbstractArray}

Returns an iterable Array of the columns in an OddFrame.

example

od = OddFrame(:A => [1, 2, 3], :B => [5, 8, 3])
columns(od)

[[1, 2, 3], [5, 8, 3]]

OddFrames.labels — Method

Interface

Iteration

labels(fg::OddFrameContainer) -> ::Vector{Symbol}

This quickly returns an iterable Array of the frames' labels in a framegroup.

example

labels = [:od1, :od2]
fg = FrameGroup(OddFrame(:y => [5, 8, 2]), OddFrame(:x => [5, 15]),
labels = labels)
labels(fg)

[:od1, :od2]

Member Functions

OddFrames have member functions. These functions are provided as fields of the OddFrame. That being said, you can access them how you normally would with fields. The advantage to this is that we can mutate the OddFrame internally. For example, here is a call of head:

od = OddFrame(:A => [5, 10, 15, 20])
od.head(5)

Non-mutating Member Functions

These functions can return a mutated frame, but will not mutate the data inside of the OddFrame.

Introspective

head

The od.head() member function will show x number of rows of the OddFrame.

head(x::Int64; html = :show) = _head(labels, columns, types, x,
html = html)
head() = _head(labels, columns, types, 5)

dtype

od.dtype() reveals the data-type of a given column by symbol index.

dtype(x::Symbol) = typeof(types[findall(x->x == x,
                        labels)[1]][1])

describe

od.describe() describes each column, or a particular column's features.

describe() = _describe(labels, columns, types)
describe(col::Symbol) = _describe(symb, labels, columns, types)

Filtering

not

Not excludes a symbol, range, or integer position from a column.

not(ls::Symbol ...) = _not(ls, labels, columns)
not(ls::UnitRange ...) = _not(ls, labels, columns)
not(ls::Int64 ...) = _not(ls, labels, columns)

not (generic function with 3 methods)

only

Only excludes all of the values except for the one provided.

only(ls::Symbol ...) = _only(ls, labels, columns)
only(ls::UnitRange ...) = _only(ls, labels, columns)
only(ls::Int64 ...) = _only(ls, labels, columns)

only (generic function with 3 methods)

Operations

apply

Apply will apply f to the columns.

apply(f::Function) = apply(f, labels, columns)

apply (generic function with 1 method)

Mutating Member Functions

These functions will mutate your type! Be careful!

drop!

The drop! function can drop a column, row, or values that meet certain parameters